Added fading to Ogg music mod
Freeware Cave Story seems to have a bug with its Org music: when the music
is meant to fade, it only get quieter. This isn't the case with the later
ports. 3DS Eshop Cave Story fades to silence with its Org soundtrack, as
does CS Wii's MIDI music. This patch does the same. Who knows, I might fix
this bug in another patch.