(Ogg music) Changed location of Ogg music to the DLL folder
Now, the settings.txt 'playlist' is not the name of a playlist txt file,
but rather the name of a folder that contains both a 'playlist.txt' file,
but also the Ogg files. Entries in the 'playlist.txt' are relative to this
folder. This makes it way simpler to juggle multiple soundtracks.