(Graphics enhancement, 60 FPS) Added VSync and merged 60 FPS mod
This is a pretty big commit. Basically, borderless is now merely an
option: I repaired the original exclusive fullscreen. The reason I
brought it back was so I could add V-Sync, since I was told the 60 FPS
mod has some pretty stuttery frame output.

Combined with the remove-sprite-alignment mod, this make Cave Story
pretty smooth. This'll be fun to playtest.