(Alt. music) Added (and switched to) stb_vorbis decoder
This uses the fancy single-header version that's currently stuck in a
Pull Request. As of writing, that branch is up-to-date.

Fun fact: Nicalis used stb_vorbis in the Switch version of Cave Story+.

Also yay 100KB-smaller DLL. Maybe I'll add dr-flac next. libmodplug is
public-domain, but it's one of those kitchen-sink libs so I'm not too
tempted.