Fix weapon floating away from player in 320x240
This issue would appear when the game is set to 320x240 in DoConfig,
the 'remove_sprite_alignment' option is enabled, and 'window_upscale'
is set to anything other than 1.

The mod would pretend it was running at a non-320x240 resolution,
causing it to screw up some pixel-conversion-calculations, causing
the bug.